refactor: edit chatbot to react hook
